+++ src/Command/SlmodImportCommand.php
@@ -9,7 +9,6 @@ use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Component\Stopwatch\Stopwatch;
@@ -45,6 +44,7 @@ class SlmodImportCommand extends Command
         $server = $this->entityManager->getRepository(Server::class)->findOneByCode($serverCode);
         if (null === $server) {
             $io->error(sprintf('Server code %s not found', $serverCode));
+
             return 1;
         }