+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.1.0
+* ADD public - graph de fréquentation du serveur DCS sur 24h
+* ADD public - ajout du graph heatmap de fréquentation du serveur sur 2 semaines
+
 ### 1.0.3
 * FIX public - format des url du module CMS